Move IComponentHTMLViewFactory to Editor from Document
